move closeSidebar function to utils
