--- web/src/hooks/closeSidebar.ts
@@ -1,6 +1,0 @@
-const closeSidebar = () => {
-  const sidebarEl = document.body.querySelector(".sidebar-wrapper") as HTMLDivElement;
-  sidebarEl.style.display = "none";
-};
-
-export default closeSidebar;
